Rename _internal.h to _private.h
