+++ test/cbmc/include/http_cbmc_state.h
@@ -32,7 +32,7 @@
 #include <stdint.h>
 
 #include "core_http_client.h"
-#include "core_http_client_internal.h"
+#include "core_http_client_private.h"
 #include "http_parser.h"
 #include "transport_interface_stubs.h"
